requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
google-generativeai==0.8.4
PyMuPDF==1.25.5
Pillow==11.1.0